Support display of images in documents not located in the project root
This commit adds support for displaying images with a relative path seen
from the document, and not only from the project root. Previously mpls
would not be able to show images used in markdown files in sub
directories of the project root. e.g if your markdown file was located
in a sub directory of the project root called docs together with some
images, mpls would not find the images since they would be referenced
with either ./ or just the filenames, and thus miss the docs folder in
the path. This commit fixes this.